% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Calculate standard error.}
\usage{
se(x)
}
\arguments{
\item{x}{A number.}
}
\value{
The standard error of \code{x}.
}
\description{
Calculate standard error.
}
\examples{
se(rnorm(10))
se(rnorm(50))
}
